Refactor BitBar utils
